bugref:9436. changed uint32_t to ULONG.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@78736 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VBoxManage/VBoxManageCloud.cpp
@@ -493,8 +493,8 @@ static RTEXITCODE showCloudInstanceInfo(HandlerArg *a, int iFirst, PCLOUDCOMMONO
     CHECK_ERROR2_RET(hrc, a->virtualBox, CreateAppliance(pAppliance.asOutParam()), RTEXITCODE_FAILURE);
 
     com::SafeIfaceArray<IVirtualSystemDescription> vsdArray;
-    uint32_t requestedVSDnums = 1;
-    uint32_t newVSDnums = 0;
+    ULONG requestedVSDnums = 1;
+    ULONG newVSDnums = 0;
     CHECK_ERROR2_RET(hrc, pAppliance, CreateVirtualSystemDescriptions(requestedVSDnums, &newVSDnums), RTEXITCODE_FAILURE);
     if (requestedVSDnums != newVSDnums)
         return RTEXITCODE_FAILURE;
